+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/previewdata/PreviewMessageData.kt
@@ -2,6 +2,7 @@ package io.getstream.chat.android.compose.previewdata
 
 import com.getstream.sdk.chat.model.ModelType
 import io.getstream.chat.android.client.models.Message
+import io.getstream.chat.android.client.models.Reaction
 import java.util.Date
 
 /**
@@ -19,5 +20,12 @@ internal object PreviewMessageData {
         text = "Aenean commodo ligula eget dolor."
         createdAt = Date()
         type = ModelType.message_regular
+    }
+
+    val messageWithOwnReaction: Message = Message().apply {
+        text = "Pellentesque leo dui, finibus et nibh et, congue aliquam lectus"
+        createdAt = Date()
+        type = ModelType.message_regular
+        ownReactions = mutableListOf(Reaction(messageId = this.id, type = "haha"))
     }
 }