+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/input/MessageInputView.kt
@@ -158,6 +158,13 @@ public class MessageInputView : ConstraintLayout {
         init(attrs)
     }
 
+    /*
+    * Sets direction of MessageInputTextDirection. Default is TEXT_DIRECTION_FIRST_STRONG_LTR
+    */
+    public fun setInputTextDirection(direction: Int) {
+        binding.messageInputFieldView.binding.messageEditText.textDirection = direction
+    }
+
     private fun configInputMode(previousValue: InputMode, newValue: InputMode) {
         when (newValue) {
             is InputMode.Reply -> {